requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
pathlib
python-dotenv
sarvamai
langdetect